# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Metawall API",
    "description": "Metawall API Documentation "
  },
  "host": "127.0.0.1:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Auth",
      "description": "註冊、登入、登出 API"
    },
    {
      "name": "Users",
      "description": "使用者相關 API，包含會員資料、追蹤、被追蹤等功能"
    },
    {
      "name": "Posts",
      "description": "貼文相關 API，包含貼文、留言、貼文按讚等功能"
    },
    {
      "name": "Upload",
      "description": "上傳 API"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "Authorization",
      "description": "請在取得的 token 前補上 \"Bearer \" 再送出(須包含一空白字元)，範例：\"Bearer {your token}\""
    }
  },
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "登入",
        "description": "登入",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "登入資訊",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "Abc123456"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "註冊",
        "description": "註冊",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "註冊資訊",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "shuantisagenius"
                },
                "email": {
                  "type": "string",
                  "example": "shuantt@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "A123456789"
                },
                "confirmPassword": {
                  "type": "string",
                  "example": "A123456789"
                }
              },
              "required": [
                "name",
                "email",
                "password",
                "confirmPassword"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "取得所有會員",
        "description": "取得所有會員",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "取得特定會員",
        "description": "取得特定會員",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "會員 ID"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "刪除特定會員",
        "description": "刪除會員，需 admin 帳號權限",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "會員 ID"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users/me/profile": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "更新會員資料",
        "description": "更新會員資料",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "會員資料",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "shuantt"
                },
                "photo": {
                  "type": "string",
                  "example": "url"
                },
                "gender": {
                  "type": "string",
                  "example": "female"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users/me/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "重設密碼",
        "description": "重設密碼",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "新密碼",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "123456"
                }
              },
              "required": [
                "password"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users/me": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "刪除會員",
        "description": "刪除會員",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users/{userId}/posts": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "取得特定會員貼文",
        "description": "取得特定會員貼文",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "會員 ID"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/{userId}/followers": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "取得跟隨者名單",
        "description": "取得跟隨者名單",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "會員 ID"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/{userId}/followings": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "取得追蹤中名單",
        "description": "取得追蹤中名單",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "會員 ID"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/me/followings/{targetUserId}": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "追蹤會員",
        "description": "追蹤會員",
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "追蹤對象會員 ID"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "取消追蹤會員",
        "description": "取消追蹤會員",
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "取消追蹤對象會員 ID"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "取得所有貼文",
        "description": "取得所有貼文",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "查詢貼文官內容關鍵字",
            "required": false,
            "type": "string"
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "查詢貼文發布時間，開始時間",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "查詢貼文發布時間，結束時間",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "size",
            "in": "query",
            "description": "查詢貼文數量，上限 100 筆，預設 100 筆",
            "required": false,
            "type": "integer",
            "maximum": 100
          },
          {
            "name": "sort",
            "in": "query",
            "description": "發文時間排序方式，預設 desc 降冪排序",
            "required": false,
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ]
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "新增貼文",
        "description": "新增貼文",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "新增貼文資訊",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "這是一篇測試文章"
                },
                "image": {
                  "type": "string",
                  "example": "img-url"
                }
              },
              "required": [
                "content"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/{postId}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "取得特定貼文",
        "description": "取得特定貼文",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文 id"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "Posts"
        ],
        "summary": "更新貼文",
        "description": "更新貼文",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "更新貼文資訊",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "小智是個大渣男!"
                },
                "image": {
                  "type": "string",
                  "example": "img-url2"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "刪除特定貼文",
        "description": "刪除特定貼文",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文 id"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/{postId}/comments": {
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "新增留言",
        "description": "新增留言",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文 id"
          },
          {
            "name": "body",
            "in": "body",
            "description": "新增留言資訊",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "小智是個大渣男!"
                }
              },
              "required": [
                "content"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/{postId}/comments/{commentId}": {
      "patch": {
        "tags": [
          "Posts"
        ],
        "summary": "更新留言",
        "description": "更新留言",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文 id"
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "留言 id"
          },
          {
            "name": "body",
            "in": "body",
            "description": "更新留言資訊",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "小智是個大渣男!"
                }
              },
              "required": [
                "content"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "刪除特定留言",
        "description": "刪除特定留言",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文 id"
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "留言 id"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/{postId}/likes": {
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "貼文按讚",
        "description": "貼文按讚",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文 id"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "取消貼文按讚",
        "description": "取消貼文按讚",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文 id"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/upload/file": {
      "post": {
        "tags": [
          "Upload"
        ],
        "description": "上傳圖片",
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "description": "圖片",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "成功訊息": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "成功訊息"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "key": {
                "type": "string",
                "example": "value"
              }
            }
          }
        }
      }
    },
    "失敗訊息": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "錯誤訊息說明"
        }
      }
    }
  }
}